Cleaning up Repository
