--- Execute.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-python3 Draculog_Executor.py
